confirm delete component
